Put utility functions in utils file
