+++ restapi/views.py
@@ -19,7 +19,7 @@ from rest_framework import status
 from restapi.models import Groups, Category, Expenses, UserExpense
 from restapi.serializers import UserSerializer, CategorySerializer, ExpensesSerializer
 from restapi.custom_exception import UnauthorizedUserException
-
+from utils import aggregate, sort_by_time_stamp, response_format, transform
 
 
 def index(_request) -> Response:
@@ -180,7 +180,7 @@ class ExpensesViewSet(ModelViewSet):
 @api_view(['post'])
 @authentication_classes([])
 @permission_classes([])
-def logProcessor(request) -> Response:
+def log_processor(request) -> Response:
     data = request.data
     num_threads = data['parallelFileProcessingCount']
     log_files = data['logFiles']
@@ -197,61 +197,6 @@ def logProcessor(request) -> Response:
     response = response_format(data)
     return Response({"response":response}, status=status.HTTP_200_OK)
 
-def sort_by_time_stamp(logs) -> list:
-    data = []
-    for log in logs:
-        data.append(log.split(" "))
-    # print(data)
-    data = sorted(data, key=lambda elem: elem[1])
-    return data
-
-def response_format(raw_data) -> list:
-    response = []
-    for timestamp, data in raw_data.items():
-        entry = {'timestamp': timestamp}
-        logs = []
-        data = {k: data[k] for k in sorted(data.keys())}
-        for exception, count in data.items():
-            logs.append({'exception': exception, 'count': count})
-        entry['logs'] = logs
-        response.append(entry)
-    return response
-
-def aggregate(cleaned_logs) -> dict:
-    data = {}
-    for log in cleaned_logs:
-        [key, text] = log
-        value = data.get(key, {})
-        value[text] = value.get(text, 0)+1
-        data[key] = value
-    return data
-
-
-def transform(logs) -> list:
-    result = []
-    for log in logs:
-        [_, timestamp, text] = log
-        text = text.rstrip()
-        timestamp = datetime.utcfromtimestamp(int(int(timestamp)/1000))
-        hours, minutes = timestamp.hour, timestamp.minute
-        key = ''
-
-        if minutes >= 45:
-            if hours == 23:
-                key = "{:02d}:45-00:00".format(hours)
-            else:
-                key = "{:02d}:45-{:02d}:00".format(hours, hours+1)
-        elif minutes >= 30:
-            key = "{:02d}:30-{:02d}:45".format(hours, hours)
-        elif minutes >= 15:
-            key = "{:02d}:15-{:02d}:30".format(hours, hours)
-        else:
-            key = "{:02d}:00-{:02d}:15".format(hours, hours)
-
-        result.append([key, text])
-        print(key)
-
-    return result
 
 
 def reader(url, timeout):
